# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MedConsult API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /users/me/sync:
    post:
      operationId: syncCurrentUser
      tags: [users]
      summary: Sync Clerk user to local DB after sign-in/register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUserBody"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      parameters:
        - name: agencyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: role
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # --- SURGEONS ---
  /surgeons:
    get:
      operationId: listSurgeons
      tags: [surgeons]
      summary: List surgeons
      parameters:
        - name: agencyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Surgeons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Surgeon"
    post:
      operationId: createSurgeon
      tags: [surgeons]
      summary: Create surgeon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurgeonBody"
      responses:
        "201":
          description: Created surgeon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Surgeon"
  /surgeons/{id}:
    get:
      operationId: getSurgeon
      tags: [surgeons]
      summary: Get surgeon by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Surgeon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Surgeon"
        "404":
          description: Not found
    patch:
      operationId: updateSurgeon
      tags: [surgeons]
      summary: Update surgeon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSurgeonBody"
      responses:
        "200":
          description: Updated surgeon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Surgeon"
    delete:
      operationId: deleteSurgeon
      tags: [surgeons]
      summary: Delete surgeon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- PROCEDURES ---
  /surgeons/{surgeonId}/procedures:
    get:
      operationId: listProcedures
      tags: [procedures]
      summary: List procedures for a surgeon
      parameters:
        - name: surgeonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Procedures list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Procedure"
    post:
      operationId: createProcedure
      tags: [procedures]
      summary: Add procedure to surgeon
      parameters:
        - name: surgeonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcedureBody"
      responses:
        "201":
          description: Created procedure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Procedure"
  /surgeons/{surgeonId}/procedures/{id}:
    patch:
      operationId: updateProcedure
      tags: [procedures]
      summary: Update procedure
      parameters:
        - name: surgeonId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProcedureBody"
      responses:
        "200":
          description: Updated procedure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Procedure"
    delete:
      operationId: deleteProcedure
      tags: [procedures]
      summary: Delete procedure
      parameters:
        - name: surgeonId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- EVENTS ---
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
          enum: [app_owner, admin, surgeon, customer]
        agencyId:
          type: ["integer", "null"]
        surgeonId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - role
        - createdAt
        - updatedAt
    SyncUserBody:
      type: object
      properties:
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
      required:
        - clerkId
        - email
    UpdateUserBody:
      type: object
      properties:
        role:
          type: ["string", "null"]
          enum: [app_owner, admin, surgeon, customer, "null"]
        agencyId:
          type: ["integer", "null"]
    Surgeon:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        agencyId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agencyId
        - firstName
        - lastName
        - email
        - createdAt
        - updatedAt
    CreateSurgeonBody:
      type: object
      properties:
        agencyId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
      required:
        - agencyId
        - firstName
        - lastName
        - email
    UpdateSurgeonBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
    Procedure:
      type: object
      properties:
        id:
          type: integer
        surgeonId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - surgeonId
        - name
        - createdAt
    CreateProcedureBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
      required:
        - name
    UpdateProcedureBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
...